pyproject.toml
[project]
name = "web-hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "autopep8>=2.3.2",
    "fastapi>=0.116.1",
    "fastapi-cache2[redis]>=0.2.2",
    "faststream[rabbit]>=0.5.48",
    "itsdangerous>=2.2.0",
    "passlib>=1.7.4",
    "phonenumbers>=9.0.13",
    "pydantic-extra-types>=2.10.5",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "pylint>=3.3.8",
    "python-jose>=3.5.0",
    "sqladmin>=0.21.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
